+++ root_orchestrator/resource-abstractor/services/hook_service.py
@@ -1,8 +1,11 @@
+import json
 import logging
 import os
 import threading
+from functools import wraps
 
 from db import hooks_db
+from flask import request
 from requests import exceptions, post
 
 RESPONSE_TIMEOUT = os.environ.get("HOOK_REQUEST_TIMEOUT", 5)
@@ -64,3 +67,86 @@ def before_create(entity_name, data):
 
 def before_update(entity_name, data):
     return process_sync_hook(entity_name, hooks_db.HookEventsEnum.BEFORE_UPDATE.value, data)
+
+
+def perform_create(entity_name, fn, *args):
+    data = args[-1]
+    data = before_create(entity_name, data)
+
+    args = list(args)
+    args[-1] = data
+
+    result = fn(*tuple(args))
+    after_create(entity_name, str(result.get("_id")))
+
+    return result
+
+
+def perform_update(entity_name, fn, *args):
+    data = args[-1]
+    data = before_update(entity_name, data)
+
+    # Update the last argument with the new data
+    args = list(args)
+    args[-1] = data
+
+    result = fn(*tuple(args))
+    after_update(entity_name, str(result.get("_id")))
+
+    return result
+
+
+# put currently is not supported
+api_hooks_map = {
+    "post": (before_create, after_create),
+    "patch": (before_update, after_update),
+    "delete": (None, after_delete),
+}
+
+
+# doesn't work well with @arguments decorator
+# this decorator has to be the closest to the method
+def before_after_hook(entity_name, with_param_id=None):
+    def decorator(fn):
+        @wraps(fn)
+        def wrapper(*args, **kwargs):
+            method_name = fn.__name__
+            if method_name not in api_hooks_map.keys():
+                raise ValueError(f"Method {method_name} not supported")
+
+            # incase this was wrapped by @arguments
+            data = args[1] if len(args) > 1 else request.json
+
+            before_fn = api_hooks_map[method_name][0]
+            after_fn = api_hooks_map[method_name][1]
+
+            entity_id = kwargs.get(with_param_id) if with_param_id else None
+
+            if before_fn:
+                if entity_id:
+                    data["_id"] = entity_id
+
+                data = before_fn(entity_name, data)
+
+            args = list(args)
+            if data and len(args) > 1:
+                args[1] = data
+            elif data:
+                args.append(data)
+
+            result = fn(*tuple(args), **kwargs)
+            result_id = result["_id"] if isinstance(result, dict) else None
+
+            if result_id is None:
+                try:
+                    result_id = json.loads(result).get("_id")
+                except json.JSONDecodeError:
+                    pass
+
+            after_fn(entity_name, result_id)
+
+            return result
+
+        return wrapper
+
+    return decorator
